requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.15.2
docx2pdf==0.1.8
psycopg2-binary==2.9.10
pywin32==308
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1